+++ pulsar/async/__init__.py
@@ -1,6 +1,7 @@
 from .access import *
 from .defer import *
 from .iostream import *
+from .transports import *
 from .mailbox import *
 from .eventloop import *
 from .proxy import *